+++ src/main/java/racingcar/model/RandomUtil.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.model;
 
 import java.util.Random;
